requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
pymysql==1.0.2
openai==0.27.0
gunicorn==20.1.0
mysql-connector-python==8.0.27
mysqlclient==2.1.0